requirements.txt
pyaudio==0.2.14
webrtcvad==2.0.10
funasr==1.1.14
modelscope==1.18.1
openai==1.51.0
python-dotenv==1.0.1
simpleaudio==1.0.4
edge-tts==6.1.14
